Restore approved minutes integer cast

+++ app/(app)/dash/makers/page.tsx
@@ -39,7 +39,7 @@ export default async function MakersPage({
       approvedMinutes: sql<number>`coalesce((
         select sum(${projects.approvedMinutes}) from ${projects}
         where ${projects.userSub} = ${users.sub} and ${projects.decision} = 'approved'
-      ), 0)`,
+      ), 0)::int`,
       beans: sql<number>`coalesce((
         select sum(${beansLedger.delta}) from ${beansLedger}
         where ${beansLedger.userSub} = ${users.sub}
